requirements.txt
Flask
gunicorn
requests
bs4
openai
python-dotenv
oauth2client
gspread